nurikabe improvements / code cleanup
The squares are now indexed so that the solution array looks like how the solution grid will actually look. Also deleted unnecessary code and renamed the html file to nurikabe. Also added the rules of the puzzle in comments at the top of the code in case anyone wants to test the puzzle out