Nurikabe hint button
When the hint button is clicked, you are shown four correct square states in the puzzle, and these squares can no longer be clicked. Hint squares are 2x2 and cannot overlap. If there is nowhere to put a hint square, a popup tells you as such and then fades out.